Use test directory structure in tracpro.polls
